requirements.txt
numpy<2
fastapi
uvicorn[standard]
transformers>=4.43.0
safetensors
torch==2.0.1+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
python-dotenv
